Added persistent triangulations and option for restarting meshes.
It turns out that it is not possible to simply save and restore a
refined mesh - none of the formats allow storing information about
hanging nodes, and there's also a problem with saving
and restoring e.g. material ID's. Since the cell order is also not
retained, it is also not possible to store this information
separately.

However, using the PersistentTriangulation class, storing the
refinement sequence from a given coarse mesh, seems to work well.